+++ backend/test/unit/oauthHelpers.test.ts
@@ -1,8 +1,10 @@
-import { describe, expect, it } from '@jest/globals';
+import { describe, expect, it, jest } from '@jest/globals';
 import {
   buildEndpointURL,
   buildRedirectUrl
 } from '../../src/utils/helpers/oauthHelpers';
+
+jest.mock('../../src/config', () => {});
 
 describe('buildEndpointURL', () => {
   it('should construct the full callback URL correctly', () => {
